refactor: extract YouTube iframe API loader
