+++ src/components/recorder/youtube-reference.tsx
@@ -12,6 +12,7 @@ import {
   RecorderRuntime,
   type ReferenceVideoState,
 } from "../../lib/recorder/runtime";
+import { loadYouTubeApi, type YouTubePlayerApi } from "../../lib/youtube";
 import { listenPointerDrag } from "../../utils/pointer-drag";
 import { Button } from "../ui/button";
 import { FloatingPanel } from "../ui/floating-panel";
@@ -96,59 +97,6 @@ export function ReferenceVideoPanel({
       />
     </FloatingPanel>
   );
-}
-
-interface YouTubePlayerApi {
-  playVideo(): void;
-  pauseVideo(): void;
-  mute(): void;
-  unMute(): void;
-  seekTo(seconds: number, allowSeekAhead: boolean): void;
-  getDuration(): number;
-  getVideoData(): { title?: string };
-  destroy(): void;
-}
-
-interface YouTubeApi {
-  Player: new (
-    element: HTMLElement,
-    options: {
-      videoId: string;
-      host?: string;
-      playerVars: Record<string, number>;
-      events: {
-        onReady: () => void;
-        onError: () => void;
-        onStateChange: (event: { data: number }) => void;
-      };
-    },
-  ) => YouTubePlayerApi;
-}
-
-declare global {
-  interface Window {
-    YT?: YouTubeApi;
-    onYouTubeIframeAPIReady?: () => void;
-  }
-}
-
-let youtubeApiPromise: Promise<YouTubeApi> | undefined;
-
-function loadYouTubeApi(): Promise<YouTubeApi> {
-  if (window.YT) {
-    return Promise.resolve(window.YT);
-  }
-  if (!youtubeApiPromise) {
-    youtubeApiPromise = new Promise((resolve, reject) => {
-      const script = document.createElement("script");
-      script.src = "https://www.youtube.com/iframe_api";
-      script.onerror = () =>
-        reject(new Error("Could not load YouTube player."));
-      window.onYouTubeIframeAPIReady = () => resolve(window.YT!);
-      document.head.appendChild(script);
-    });
-  }
-  return youtubeApiPromise;
 }
 
 function YouTubeReferenceSetup({
